Do not round corners on tall selection rectangles
When performing the smart select animation, some of the selection
rectangles may be taller than they are wide (e.g. due to a very large
line-height). Since in these cases we cannot properly round the corners
and the effect looks off, the agreed upon action is to not round the
corners of these selection rectangles at all.

Test: manual - run SmartSelectSprite on a manually constructed set of
rectangles and verify that it behaves correctly

Change-Id: Ifd02005e3d2709736c34f6d07671a4eb7a9557db