use simOperator name in the notification
today we are using getNetworkOperator for carrier's name in the
notification which might be misleading in cases users romaing on
different network.
as today there is no reliable way to get carrier's name, we will use
getSimOperator and if that returns empty string, fallback to 
networkOperatorName. In the long term, we can use better wording if 
simOperatorName is unknown.

Bug:65251050
Test: Manual
Change-Id: I6ffc37f8cc8e686d2e884283be2006cf704e8c64