Watchdog: Remove kernel thread dumps (aka Watchdog_N).
It's not clear that this is useful any more - Java threads and native
threads can now be unwound through kernel frames so a regular watchdog
bite will contain kernel traces for threads attached to the runtime.

Note that the code being removed would dump traces for *all* threads,
even those not attached to the runtime but it's unclear whether we're
gaining extra given that they will be useful only if a watchdog
monitored thread is blocked on some other (non-attached) thread
performing work.

Moreover, the warning messages from Watchdog_N on user builds seems to
be causing confusion :
09-07 22:41:10.005  1000   950  1557 E Watchdog_N: Unable to open stack of tid 1474 : 13 (Permission denied)
09-07 22:41:10.006  1000   950  1557 E Watchdog_N: Unable to open stack of tid 1475 : 13 (Permission denied)

Bug: 65475549
Test: make & flash
Change-Id: Iad5022d67fecdab5f1130cf04da12b425c63ca7a