Apply PGO for all modules in hwui
Bug: http://b/65598278

Since hwui source files get statically linked into hwuimicro and
hwuimacro, we need to enable PGO for them (to instrument and collect
profiles).  For consistency, apply PGO to all hwui modules (including
the unit tests) so we test the actual build configuration for libhwui.

For now, we don't use any profile so this CL affects the build only when
instrumentation is turned on using 'ANDROID_PGO_INSTRUMENT=hwui'.

Test: Build with ANDROID_PGO_INSTRUMENT=hwui, run hwuimacro and
hwuimicro and verify profile generation

Change-Id: I9d35953d544e0ccc180eb164ff53466239194c0f