Allow a11y shortcut to toggle framework features
Enabling color inversion and color correction initially.

Not showing toast for the framework features, as both are
instantly visible changes and the extra reminder seems like
it will get in the way.

Moving AccessibilityShortcutController to a place where
it can be seen by Settings, so Settings can learn what
framework features to offer the user. Moving tests for that
class to match.

Currently don't have icons for the two framework features.
They will be added in a future CL once I get them.

Also tweaking the warning dialog to include summary
information, if we have it, for the target service.

Bug: 34621067
Bug: 36368472
Test: Adding unit tests for framework features
Change-Id: I32a10989db1c9ad9bf22aae9ad405771b789bc6f